Unify qrstart config generation
