+++ config-model/src/main/java/com/yahoo/vespa/model/admin/LogserverContainer.java
@@ -15,7 +15,7 @@ public class LogserverContainer extends Container {
     public LogserverContainer(AbstractConfigProducer parent) {
         super(parent, "" + 0, 0);
         addComponent(new AccessLogComponent(AccessLogComponent.AccessLogType.jsonAccessLog, ((LogserverContainerCluster) parent).getName(), true));
-        appendJvmOptions("-Xms32m -Xmx512m");
+        appendJvmOptions("-Xms32m");
     }
 
     @Override
