+++ config-model/src/main/java/com/yahoo/vespa/model/admin/LogserverContainerCluster.java
@@ -4,7 +4,6 @@ package com.yahoo.vespa.model.admin;
 import com.yahoo.config.model.deploy.DeployState;
 import com.yahoo.config.model.producer.AbstractConfigProducer;
 import com.yahoo.container.handler.ThreadpoolConfig;
-import com.yahoo.search.config.QrStartConfig;
 import com.yahoo.vespa.model.container.ContainerCluster;
 import com.yahoo.vespa.model.container.component.Handler;
 
@@ -22,14 +21,6 @@ public class LogserverContainerCluster extends ContainerCluster<LogserverContain
 
     @Override
     protected void doPrepare(DeployState deployState) { }
-
-    // Switch off verbose:gc, it's very noisy when Xms < Xmx
-    @Override
-    public void getConfig(QrStartConfig.Builder builder) {
-        super.getConfig(builder);
-        // This takes effect via vespa-start-container-daemon:configure_gcopts
-        builder.jvm.verbosegc(false);
-    }
 
     @Override
     public void getConfig(ThreadpoolConfig.Builder builder) {
