+++ config-model/src/main/java/com/yahoo/vespa/model/container/ContainerCluster.java
@@ -37,6 +37,7 @@ import com.yahoo.vespa.configdefinition.IlscriptsConfig;
 import com.yahoo.vespa.model.PortsMeta;
 import com.yahoo.vespa.model.Service;
 import com.yahoo.vespa.model.admin.monitoring.Monitoring;
+import com.yahoo.vespa.model.application.validation.RestartConfigs;
 import com.yahoo.vespa.model.clients.ContainerDocumentApi;
 import com.yahoo.vespa.model.container.component.AccessLogComponent;
 import com.yahoo.vespa.model.container.component.Component;
@@ -78,6 +79,7 @@ import static com.yahoo.container.core.BundleLoaderProperties.DISK_BUNDLE_PREFIX
  * @author Einar M R Rosenvinge
  * @author Tony Vaagenes
  */
+@RestartConfigs({QrStartConfig.class})
 public abstract class ContainerCluster<CONTAINER extends Container>
         extends AbstractConfigProducer<AbstractConfigProducer<?>>
         implements
@@ -484,13 +486,12 @@ public abstract class ContainerCluster<CONTAINER extends Container>
 
     @Override
     public void getConfig(QrStartConfig.Builder builder) {
-        QrStartConfig.Jvm.Builder jvmBuilder = builder.jvm;
-        if (getMemoryPercentage().isPresent()) {
-            jvmBuilder.heapSizeAsPercentageOfPhysicalMemory(getMemoryPercentage().get());
-        } else if (isHostedVespa()) {
-            jvmBuilder.heapSizeAsPercentageOfPhysicalMemory(getHostClusterId().isPresent() ? 17 : 60);
-        }
-        jvmBuilder.gcopts(Objects.requireNonNullElse(jvmGCOptions, G1GC));
+        builder.jvm
+                .verbosegc(false)
+                .availableProcessors(2)
+                .heapsize(512)
+                .heapSizeAsPercentageOfPhysicalMemory(0)
+                .gcopts(Objects.requireNonNullElse(jvmGCOptions, G1GC));
         if (environmentVars != null) {
             builder.qrs.env(environmentVars);
         }
